refactor(core): reorganization of SmartHopper.Core
