+++ .windsurf/rules/ai-tool-conventions.md
@@ -4,7 +4,7 @@ globs: **/SmartHopper.Core.Grasshopper/Tools/*.cs
 ---
 
 # AI Tool conventions
-- Implement IAIToolProvider in SmartHopper.Core.Grasshopper.Tools
+- Implement IAIToolProvider in SmartHopper.Core.Grasshopper.AITools
 - File name: scope_action.cs
 - Define AITool metadata (Name, Description, schema)
 - Auto-discover via AIToolManager
