+++ src/SmartHopper.Components/List/AIListFilter.cs
@@ -22,7 +22,7 @@ using Newtonsoft.Json.Linq;
 using SmartHopper.Components.Properties;
 using SmartHopper.Core.ComponentBase;
 using SmartHopper.Core.DataTree;
-using SmartHopper.Core.Grasshopper.Tools;
+using SmartHopper.Core.Grasshopper.Utils;
 
 namespace SmartHopper.Components.List
 {
