+++ src/SmartHopper.Components/Text/AITextEvaluate.cs
@@ -21,7 +21,6 @@ using Newtonsoft.Json.Linq;
 using SmartHopper.Components.Properties;
 using SmartHopper.Core.ComponentBase;
 using SmartHopper.Core.DataTree;
-using SmartHopper.Core.Grasshopper.Tools;
 using CommonDrawing = System.Drawing;
 
 namespace SmartHopper.Components.Text
