+++ src/SmartHopper.Core.Grasshopper/AITools/list_evaluate.cs
@@ -18,10 +18,11 @@ using Grasshopper.Kernel.Types;
 using Newtonsoft.Json.Linq;
 using SmartHopper.Infrastructure.Interfaces;
 using SmartHopper.Infrastructure.Models;
-using SmartHopper.Core.AI;
 using SmartHopper.Core.Grasshopper.Models;
+using SmartHopper.Core.Grasshopper.Utils;
+using SmartHopper.Core.Messaging;
 
-namespace SmartHopper.Core.Grasshopper.Tools
+namespace SmartHopper.Core.Grasshopper.AITools
 {
     /// <summary>
     /// Contains tools for list analysis and manipulation using AI.
