+++ src/SmartHopper.Core.Grasshopper/AITools/text_evaluate.cs
@@ -15,12 +15,13 @@ using System.Threading.Tasks;
 using Grasshopper.Kernel;
 using Grasshopper.Kernel.Types;
 using Newtonsoft.Json.Linq;
+using SmartHopper.Core.Grasshopper.Models;
+using SmartHopper.Core.Grasshopper.Utils;
+using SmartHopper.Core.Messaging;
 using SmartHopper.Infrastructure.Interfaces;
 using SmartHopper.Infrastructure.Models;
-using SmartHopper.Core.AI;
-using SmartHopper.Core.Grasshopper.Models;
 
-namespace SmartHopper.Core.Grasshopper.Tools
+namespace SmartHopper.Core.Grasshopper.AITools
 {
     /// <summary>
     /// Contains tools for text analysis and manipulation using AI.
