+++ src/SmartHopper.Core.Grasshopper/AITools/web_rhino_forum_search.cs
@@ -17,7 +17,7 @@ using Newtonsoft.Json.Linq;
 using SmartHopper.Infrastructure.Interfaces;
 using SmartHopper.Infrastructure.Models;
 
-namespace SmartHopper.Core.Grasshopper.Tools
+namespace SmartHopper.Core.Grasshopper.AITools
 {
     /// <summary>
     /// Provides AI tools for fetching webpage text content,
