+++ src/SmartHopper.Core/UI/Chat/Resources/css/chat-styles.css
