+++ src/SmartHopper.Core/UI/Chat/Resources/js/chat-script.js
