+++ src/SmartHopper.Core/UI/Chat/Resources/js/mathjax/tex-mml-chtml.js
