+++ src/SmartHopper.Core/UI/Chat/Resources/templates/chat-template.html
