+++ src/SmartHopper.Core/UI/Chat/Resources/templates/error-template.html
