+++ src/SmartHopper.Core/UI/Chat/Resources/templates/message-template.html
